[UPD] Reworking data structures

+++ heuristic.go
@@ -1,16 +1,16 @@
 package main
 
 // selectedHeuristic represents the heuristic selected by the end user
-var selectedHeuristic func(state nodeState) int
+var selectedHeuristic func(state nodeState) float64
 
 // ManhattanHeuristic calculates the distance between each piece
 // and its final position
-func ManhattanHeuristic(state nodeState) int {
+func ManhattanHeuristic(state nodeState) float64 {
 	return 3
 }
 
 // MisplacedHeuristic calculates the number of misplaced pieces
-func MisplacedHeuristic(state nodeState) int {
+func MisplacedHeuristic(state nodeState) float64 {
 	misplacedElems := 0
 
 	for y := range state {
@@ -21,11 +21,11 @@ func MisplacedHeuristic(state nodeState) int {
 		}
 	}
 
-	return misplacedElems
+	return float64(misplacedElems)
 }
 
 // LinearConflictHeuristic is the ManhattanHeuristic ponderated by
 // the conflict between some pieces and their final destination
-func LinearConflictHeuristic(state nodeState) int {
+func LinearConflictHeuristic(state nodeState) float64 {
 	return 1
 }
